REFACTOR-#6: Replace qgrid from file names and css classes
