+++ js/src/spreadsheet.datefilter.js
@@ -1,12 +1,12 @@
 var $ = require('jquery');
-var filter_base = require('./qgrid.filterbase.js');
+var filter_base = require('./spreadsheet.filterbase.js');
 
 class DateFilter extends filter_base.FilterBase {
 
   get_filter_html() {
     return `
       <div class='date-range-filter grid-filter qgrid-dropdown-menu'>
-        <h3 class='qgrid-popover-title'>
+        <h3 class='spreadsheet-popover-title'>
           <div class='dropdown-title'>Filter by ${this.field}</div>
           <i class='fa fa-times icon-remove close-button'></i>
         </h3>
