+++ js/src/spreadsheet.textfilter.js
@@ -1,13 +1,13 @@
 var $ = require('jquery');
 var _ = require('underscore');
-var filter_base = require('./qgrid.filterbase.js');
+var filter_base = require('./spreadsheet.filterbase.js');
 
 class TextFilter extends filter_base.FilterBase {
 
   get_filter_html() {
     return `
       <div class='text-filter grid-filter qgrid-dropdown-menu'>
-        <h3 class='qgrid-popover-title'>
+        <h3 class='spreadsheet-popover-title'>
           <div class='dropdown-title'>Filter by ${this.field}</div>
           <i class='fa fa-times icon-remove close-button'></i>
         </h3>
